requirements.txt
elasticsearch==5.5.3
PyYAML==5.1
urllib3==1.24.1
